==> research/06_predictions.ipynb <==
import os

%pwd

os.chdir('../')

%pwd

from dataclasses import dataclass
from pathlib import Path


@dataclass(frozen=True)
class PredictionsConfig:
    root_dir: Path
    test_data_path: Path
    threshold_path: Path
    class_model_path: Path
    reg_model_path: Path
    potential_customers_data_path: Path
    potential_customers_with_predictions_data_path: Path


from AutoInsurance.constants import *
from AutoInsurance.utils.common import read_yaml, create_directories, save_json

class ConfigurationManager:
    def __init__(
        self,
        config_filepath = CONFIG_FILE_PATH,
        params_filepath = PARAMS_FILE_PATH,
        schema_filepath = SCHEMA_FILE_PATH):

        self.config = read_yaml(config_filepath)
        self.params = read_yaml(params_filepath)
        self.schema = read_yaml(schema_filepath)

        create_directories([self.config.artifacts_root])

    def get_predictions_config(self) -> PredictionsConfig:
        config = self.config.predictions

        create_directories([config.root_dir])
        predictions_config =  PredictionsConfig(
            root_dir = Path(config.root_dir),
            threshold_path = Path(config.threshold_path),
            test_data_path = Path(config.test_data_path),
            class_model_path = Path(config.class_model_path),
            reg_model_path = Path(config.reg_model_path),
            potential_customers_data_path = Path(config.potential_customers_data_path),
            potential_customers_with_predictions_data_path = Path(config.potential_customers_with_predictions_data_path)
        )
    
        return predictions_config

import joblib 
import pandas as pd
from AutoInsurance.utils.common import load_json
import numpy as np
from AutoInsurance.utils.common import logger

class Predictions:
    def __init__(self, config: PredictionsConfig):
        self.config = config
    

    def get_predictions(self):

        logger.info(f"class_model_path: {self.config.class_model_path}")
        logger.info(f"reg_model_path: {self.config.reg_model_path}")
        logger.info(f"test_data_path: {self.config.test_data_path}")
        logger.info(f"threshold_path: {self.config.threshold_path}")
        logger.info(f"potential_customers_data_path: {self.config.potential_customers_data_path}")


        class_model = joblib.load(Path(self.config.class_model_path))
        reg_model = joblib.load(Path(self.config.reg_model_path))

        test_data = pd.read_csv(Path(self.config.test_data_path))
        class_predictions_probs = class_model.predict_proba(test_data)[:, 1]

        metrics = load_json(Path(self.config.threshold_path))
        class_predictions = (class_predictions_probs >= metrics.optimal_threshold).astype(int)

        reg_predictions = reg_model.predict(test_data)
        reg_predictions = np.expm1(reg_predictions)

        potential_customers_data = pd.read_csv(Path(self.config.potential_customers_data_path))
        potential_customers_data['claim_probability'] = class_predictions_probs
        potential_customers_data['claim'] = class_predictions
        potential_customers_data['claim_amount'] = reg_predictions

        potential_customers_data.to_csv(Path(self.config.potential_customers_with_predictions_data_path), index = False)


try:
    config = ConfigurationManager()
    predictions_config = config.get_predictions_config()
    predictions_config = Predictions(config = predictions_config)
    predictions_config.get_predictions()

except Exception as e:
    raise e
