==> research/08_user_app_backend testing.ipynb <==
import os

%pwd

os.chdir('../')

%pwd

import pandas as pd
import numpy as np

from dataclasses import dataclass
from pathlib import Path


@dataclass(frozen=True)
class UserAppConfig:
    root_dir: Path
    risk_profiles_path: Path
    class_model_path: Path
    reg_model_path: Path
    test_data_path: Path
    scaler_path: Path
    params: dict

from AutoInsurance.constants import *
from AutoInsurance.utils.common import read_yaml, create_directories, save_json

class ConfigurationManager:
    def __init__(
        self,
        config_filepath = CONFIG_FILE_PATH,
        params_filepath = PARAMS_FILE_PATH,
        schema_filepath = SCHEMA_FILE_PATH):

        self.config = read_yaml(config_filepath)
        self.params = read_yaml(params_filepath)
        self.schema = read_yaml(schema_filepath)

        create_directories([self.config.artifacts_root])

    def get_user_app_config(self) -> UserAppConfig:
        config = self.config.user_app
        params = self.params.RiskProfiles

        create_directories([config.root_dir])

        user_app_config = UserAppConfig(
            root_dir= Path(config.root_dir),
            test_data_path= Path(config.test_data_path),
            risk_profiles_path= Path(config.risk_profiles_path),
            class_model_path= Path(config.class_model_path),
            reg_model_path= Path(config.reg_model_path),
            scaler_path= Path(config.scaler_path),
            params= params
        )

        return user_app_config       

import pandas as pd
import joblib

class RiskProfileModel:
    def __init__(self, config: UserAppConfig):
        self.config = config
        self.params = config.params

    # def load_data(self):
    #     return pd.read_csv(Path(self.config.risk_profiles_path))

    def transform_user_data_to_df(self, user_data, columns, dtypes):
        data = {col: np.zeros(1, dtype=dt) if dt == 'float64' else np.zeros(1, dtype=bool) for col, dt in dtypes.items()}
        df = pd.DataFrame(data)
        
        if 'credit_score' in user_data:
            df['credit_score'] = user_data['credit_score']
        if 'traffic_index' in user_data:
            df['traffic_index'] = user_data['traffic_index']
        if 'veh_value' in user_data:
            df['veh_value'] = user_data['veh_value']
        
        for key, value in user_data.items():
            if key in ['gender', 'area', 'veh_body', 'agecat', 'veh_age']:
                column_name = f'{key}_{value}'
                if column_name in df.columns:
                    df[column_name] = True
        
        return df
    
    def predict(self, user_data, columns, dtypes):
        df = self.transform_user_data_to_df(user_data, columns, dtypes)
        class_predictions_probs = joblib.load(Path(self.config.class_model_path)).predict_proba(df)[:, 1]
        reg_predictions = joblib.load(Path(self.config.reg_model_path)).predict(df)
        reg_predictions = np.expm1(reg_predictions)
        
        claim_likelihood = class_predictions_probs[0]
        claim_amount = reg_predictions[0]
        
        return claim_likelihood, claim_amount
    
    def normalize_predictions(self, claim_likelihood, claim_amount):

        features = pd.DataFrame([[claim_likelihood, claim_amount]], columns=['claim_probability', 'claim_amount'])
        norm_values =  joblib.load(Path(self.config.scaler_path)).transform(features)
        normalized_claim_likelihood = norm_values[0, 0]
        normalized_claim_amount = norm_values[0, 1]
        
        return normalized_claim_likelihood, normalized_claim_amount
    
    def classify_risk(self, normalized_claim_likelihood, normalized_claim_amount, claim_amount):

        risk_profiles_df = pd.read_csv(Path(self.config.risk_profiles_path))

        quantiles_prob = risk_profiles_df['claim_probability'].quantile(self.params['claim_probability_thresholds'])
        if normalized_claim_likelihood <= quantiles_prob.iloc[0]:
            risk_profile_probability = 'Low'
        elif normalized_claim_likelihood <= quantiles_prob.iloc[1]:
            risk_profile_probability = 'Medium'
        else:
            risk_profile_probability = 'High'
        
        if claim_amount == 0:
            risk_profile_cost = 'No Claim'
        else:
            quantiles_cost = risk_profiles_df.loc[risk_profiles_df['claim_amount'] > 0, 'claim_amount'].quantile(self.params['claim_amount_thresholds'])
            if normalized_claim_amount <= quantiles_cost.iloc[0]:
                risk_profile_cost = 'Low'
            elif normalized_claim_amount <= quantiles_cost.iloc[1]:
                risk_profile_cost = 'Medium'
            else:
                risk_profile_cost = 'High'

        weights_probability = self.params['weights_probability']
        weights_cost = self.params['weights_cost']
        
        weighted_probability_score = normalized_claim_likelihood * weights_probability[risk_profile_probability]
        weighted_cost_score = normalized_claim_amount * weights_cost[risk_profile_cost]
        dynamic_combined_risk_score = weighted_probability_score + weighted_cost_score
        
        quantiles_risk = risk_profiles_df['dynamic_combined_risk_score'].quantile(self.params['risk_score_thresholds'])
        if dynamic_combined_risk_score <= quantiles_risk.iloc[0]:
            risk_group = 'Low Risk'
        elif dynamic_combined_risk_score <= quantiles_risk.iloc[1]:
            risk_group = 'Medium Risk'
        else:
            risk_group = 'High Risk'
        
        return {
            'claim_likelihood': claim_likelihood,
            'claim_amount': claim_amount,
            'normalized_claim_likelihood': normalized_claim_likelihood,
            'normalized_claim_amount': normalized_claim_amount,
            'risk_profile_probability': risk_profile_probability,
            'risk_profile_cost': risk_profile_cost,
            'dynamic_combined_risk_score': dynamic_combined_risk_score,
            'risk_group': risk_group
        }


config = ConfigurationManager()
user_app_config = config.get_user_app_config()
risk_profile_model = RiskProfileModel(config = user_app_config)

test = pd.read_csv(user_app_config.test_data_path)
columns = test.columns
dtypes = test.dtypes

def user_input_features():
    data = {
        'gender': 'F',
        'agecat': '2',
        'credit score': 800,
        'area': 'A',
        'traffic index': 1.2,
        'vehicle age': '2',
        'vehicle body': 'SEDAN',
        'vehicle value': 0.9
        }
    return data

user_data = user_input_features()

user_data_df = pd.DataFrame(list(user_data.items()), columns=['Feature', 'Value']).astype(str)

claim_likelihood, claim_amount = risk_profile_model.predict(user_data, columns, dtypes)
normalized_claim_likelihood, normalized_claim_amount = risk_profile_model.normalize_predictions(claim_likelihood, claim_amount)
risk_profile = risk_profile_model.classify_risk(normalized_claim_likelihood, normalized_claim_amount, claim_amount)

risk_profile

# Testing

test = pd.read_csv('artifacts\data_transformation\Processed_test_data.csv')

test.columns

test.dtypes

import pandas as pd
import numpy as np

# Define the function to transform user data into the DataFrame
def transform_user_data_to_df(user_data, columns, dtypes):
    # Initialize the DataFrame with zeros and the specified data types
    data = {col: np.zeros(1, dtype=dt) if dt == 'float64' else np.zeros(1, dtype=bool) for col, dt in dtypes.items()}
    df = pd.DataFrame(data)
    
    # Fill in the continuous/numerical data
    if 'credit_score' in user_data:
        df['credit_score'] = user_data['credit_score']
    if 'traffic_index' in user_data:
        df['traffic_index'] = user_data['traffic_index']
    if 'veh_value' in user_data:
        df['veh_value'] = user_data['veh_value']
    
    # Handle categorical data by setting the relevant column to True
    for key, value in user_data.items():
        if key in ['gender', 'area', 'veh_body', 'agecat', 'veh_age']:
            column_name = f'{key}_{value}'
            if column_name in df.columns:
                df[column_name] = True
    
    # Return the resulting DataFrame
    return df

# Example usage
user_data = {
    'gender': 'F',
    'agecat': '3',
    'credit_score': 500,
    'area': 'A',
    'traffic_index': 1.9,
    'veh_age': '1',
    'veh_body': 'SEDAN',
    'veh_value': 1.5
}

# Initialize an empty DataFrame with the specified columns and data types
columns = test.columns
dtypes = test.dtypes

# Transform the user data into the DataFrame
df = transform_user_data_to_df(user_data, columns, dtypes)
print(df)
print(df.dtypes)


import joblib
from pathlib import Path 

class_model = joblib.load(Path("artifacts\model_trainer\class_model.joblib"))

reg_model = joblib.load(Path('artifacts/model_trainer/reg_model.joblib'))

class_predictions_probs = class_model.predict_proba(df)[:, 1]

class_predictions_probs[0]

reg_predictions = reg_model.predict(df)
reg_predictions = np.expm1(reg_predictions)
reg_predictions[0]

claim_likelihood = class_predictions_probs[0]
claim_amount = reg_predictions[0]

scaler = joblib.load(Path('artifacts/risk_profiles/minmax_scaler.pkl'))

predictions_df = pd.read_csv('artifacts/risk_profiles/risk_profiles.csv')

norm_values = scaler.transform([[claim_likelihood, claim_amount]])
normalized_claim_likelihood = norm_values[0, 0]
normalized_claim_amount = norm_values[0, 1]

predictions_df.columns

claim_probability_thresholds = [0.2, 0.45]
claim_amount_thresholds = [0.2, 0.45]
weights_probability = {'Low': 0.4, 'Medium': 0.5, 'High': 0.6}
weights_cost = {'No Claim': 0.3, 'Low': 0.4, 'Medium': 0.5, 'High': 0.6}
risk_score_thresholds = [0.2, 0.45]

quantiles_prob = predictions_df['claim_probability'].quantile(claim_probability_thresholds)
if normalized_claim_likelihood <= quantiles_prob.iloc[0]:
    risk_profile_probability = 'Low'
elif normalized_claim_likelihood <= quantiles_prob.iloc[1]:
    risk_profile_probability = 'Medium'
else:
    risk_profile_probability = 'High'

if claim_amount == 0:
    risk_profile_cost = 'No Claim'
else:
    quantiles_cost = predictions_df.loc[predictions_df['claim_amount'] > 0, 'claim_amount'].quantile(claim_amount_thresholds)
    if normalized_claim_amount <= quantiles_cost.iloc[0]:
        risk_profile_cost = 'Low'
    elif normalized_claim_amount <= quantiles_cost.iloc[1]:
        risk_profile_cost = 'Medium'
    else:
        risk_profile_cost = 'High'

weighted_probability_score = normalized_claim_likelihood * weights_probability[risk_profile_probability]
weighted_cost_score = normalized_claim_amount * weights_cost[risk_profile_cost]
    

dynamic_combined_risk_score = weighted_probability_score + weighted_cost_score

quantiles_risk = predictions_df['dynamic_combined_risk_score'].quantile(risk_score_thresholds)
if dynamic_combined_risk_score <= quantiles_risk.iloc[0]:
    risk_group = 'Low Risk'
elif dynamic_combined_risk_score <= quantiles_risk.iloc[1]:
    risk_group = 'Medium Risk'
else:
    risk_group = 'High Risk'
{
    'claim_likelihood': claim_likelihood,
    'claim_amount': claim_amount,
    'normalized_claim_likelihood': normalized_claim_likelihood,
    'normalized_claim_amount': normalized_claim_amount,
    'risk_profile_probability': risk_profile_probability,
    'risk_profile_cost': risk_profile_cost,
    'dynamic_combined_risk_score': dynamic_combined_risk_score,
    'risk_group': risk_group
    
}

import pandas as pd
import numpy as np
import joblib
from sklearn.preprocessing import MinMaxScaler
from pathlib import Path

class RiskProfileModel:
    def __init__(self, model_paths, scaler_path, risk_profile_path):
        self.class_model = joblib.load(Path(model_paths['class_model']))
        self.reg_model = joblib.load(Path(model_paths['reg_model']))
        self.scaler = joblib.load(Path(scaler_path))
        self.risk_profiles_df = pd.read_csv(risk_profile_path)
    
    def transform_user_data_to_df(self, user_data, columns, dtypes):
        data = {col: np.zeros(1, dtype=dt) if dt == 'float64' else np.zeros(1, dtype=bool) for col, dt in dtypes.items()}
        df = pd.DataFrame(data)
        
        if 'credit_score' in user_data:
            df['credit_score'] = user_data['credit_score']
        if 'traffic_index' in user_data:
            df['traffic_index'] = user_data['traffic_index']
        if 'veh_value' in user_data:
            df['veh_value'] = user_data['veh_value']
        
        for key, value in user_data.items():
            if key in ['gender', 'area', 'veh_body', 'agecat', 'veh_age']:
                column_name = f'{key}_{value}'
                if column_name in df.columns:
                    df[column_name] = True
        
        return df
    
    def predict(self, user_data, columns, dtypes):
        df = self.transform_user_data_to_df(user_data, columns, dtypes)
        class_predictions_probs = self.class_model.predict_proba(df)[:, 1]
        reg_predictions = self.reg_model.predict(df)
        reg_predictions = np.expm1(reg_predictions)
        
        claim_likelihood = class_predictions_probs[0]
        claim_amount = reg_predictions[0]
        
        return claim_likelihood, claim_amount
    
    def normalize_predictions(self, claim_likelihood, claim_amount):
        
        features = pd.DataFrame([[claim_likelihood, claim_amount]], columns=['claim_probability', 'claim_amount'])

        norm_values = self.scaler.transform(features)
        # norm_values = self.scaler.transform([[claim_likelihood, claim_amount]])
        normalized_claim_likelihood = norm_values[0, 0]
        normalized_claim_amount = norm_values[0, 1]
        
        return normalized_claim_likelihood, normalized_claim_amount
    
    def classify_risk(self, normalized_claim_likelihood, normalized_claim_amount, claim_amount):
        claim_probability_thresholds = [0.2, 0.45]
        claim_amount_thresholds = [0.2, 0.45]
        weights_probability = {'Low': 0.4, 'Medium': 0.5, 'High': 0.6}
        weights_cost = {'No Claim': 0.3, 'Low': 0.4, 'Medium': 0.5, 'High': 0.6}
        risk_score_thresholds = [0.2, 0.45]
        
        quantiles_prob = self.risk_profiles_df['claim_probability'].quantile(claim_probability_thresholds)
        if normalized_claim_likelihood <= quantiles_prob.iloc[0]:
            risk_profile_probability = 'Low'
        elif normalized_claim_likelihood <= quantiles_prob.iloc[1]:
            risk_profile_probability = 'Medium'
        else:
            risk_profile_probability = 'High'
        
        if claim_amount == 0:
            risk_profile_cost = 'No Claim'
        else:
            quantiles_cost = self.risk_profiles_df.loc[self.risk_profiles_df['claim_amount'] > 0, 'claim_amount'].quantile(claim_amount_thresholds)
            if normalized_claim_amount <= quantiles_cost.iloc[0]:
                risk_profile_cost = 'Low'
            elif normalized_claim_amount <= quantiles_cost.iloc[1]:
                risk_profile_cost = 'Medium'
            else:
                risk_profile_cost = 'High'
        
        weighted_probability_score = normalized_claim_likelihood * weights_probability[risk_profile_probability]
        weighted_cost_score = normalized_claim_amount * weights_cost[risk_profile_cost]
        dynamic_combined_risk_score = weighted_probability_score + weighted_cost_score
        
        quantiles_risk = self.risk_profiles_df['dynamic_combined_risk_score'].quantile(risk_score_thresholds)
        if dynamic_combined_risk_score <= quantiles_risk.iloc[0]:
            risk_group = 'Low Risk'
        elif dynamic_combined_risk_score <= quantiles_risk.iloc[1]:
            risk_group = 'Medium Risk'
        else:
            risk_group = 'High Risk'
        
        return {
            'claim_likelihood': claim_likelihood,
            'claim_amount': claim_amount,
            'normalized_claim_likelihood': normalized_claim_likelihood,
            'normalized_claim_amount': normalized_claim_amount,
            'risk_profile_probability': risk_profile_probability,
            'risk_profile_cost': risk_profile_cost,
            'dynamic_combined_risk_score': dynamic_combined_risk_score,
            'risk_group': risk_group
        }


model_paths = {
    'class_model': "artifacts/model_trainer/class_model.joblib",
    'reg_model': 'artifacts/model_trainer/reg_model.joblib'
}
scaler_path = 'artifacts/risk_profiles/minmax_scaler.pkl'
risk_profile_path = 'artifacts/risk_profiles/risk_profiles.csv'

# Initialize the model
risk_profile_model = RiskProfileModel(model_paths, scaler_path, risk_profile_path)

# Load test data to get columns and dtypes
test = pd.read_csv('artifacts/data_transformation/Processed_test_data.csv')
columns = test.columns
dtypes = test.dtypes

# Define user data
user_data = {
        'gender': 'F',
        'agecat': '2',
        'credit score': 800,
        'area': 'A',
        'traffic index': 0,
        'vehicle age': '2',
        'vehicle body': 'SEDAN',
        'vehicle value': 0
        }

# Get predictions
claim_likelihood, claim_amount = risk_profile_model.predict(user_data, columns, dtypes)

# Normalize predictions
normalized_claim_likelihood, normalized_claim_amount = risk_profile_model.normalize_predictions(claim_likelihood, claim_amount)

# Classify risk
risk_profile = risk_profile_model.classify_risk(normalized_claim_likelihood, normalized_claim_amount, claim_amount)

# Output the risk profile
print(risk_profile)


test.columns

test.veh_value.describe()
